backend/requirements.txt
Flask==1.1.1
Flask-HTTPAuth==3.3.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
Werkzeug==1.0.0
Flask-SQLAlchemy==2.4.1

